move time functions to new time.js module
